Miglioramento progettazione algoritmi item fields
